core/tracks: split off the *Track implementation from core, update main readme on versioning
