+++ README.md
@@ -8,11 +8,19 @@ it will be documented with examples, but for now, source is it.
 Scorpy is distributed under the GPLv2 license, except for the median calculation
 helper in statistics.py whose license is unknown (see source)
 
+# Versioning
+
+Internally, scorpy uses [semver](https://semver.org) compatible scheme. However,
+while Scorpy is initial development (major == 0), to guarantees about API
+stability or backward compatibility is given. Scorpy is not semver compliant at
+this point.
+
 # Scorpy library source code
 
 ## core
 
-Contains the core classes and generators
+Contains the core utilities and generators. Imports `tracks.*` automatically
+into the same namespace.
 
 ## reader
 
@@ -25,6 +33,10 @@ Contains reuseable higher level report functions (one ore more)
 ## statistics
 
 Median calculation support without numpy or python3
+
+## tracks
+
+Implementations of `core.*Track` classes. No need to pull this in separately.
 
 ## vcd
